+++ src/Journalist.EventStore/Streams/IEventStreamInteractionEntity.cs
@@ -1,0 +1,11 @@
+namespace Journalist.EventStore.Streams
+{
+    public interface IEventStreamInteractionEntity
+    {
+        string StreamName { get; }
+
+        StreamVersion StreamVersion { get; }
+
+        bool IsClosed { get; }
+    }
+}